Add support for multiple chants. but
But it is not perfact.

IT NEEDS UPDATE.

The problem of this method is:
You need to add chant chapters as mush as you want or it crashes.
If there is only one chant, the other chapter will be left
as blank(//)
It is a waste that you have to add extra chapters like # 입당성가1 and